Move ZoneUtil to internal utils
Zone grammar validation is an implementation detail of ServiceIdentifier,
so it lives in org.kinotic.core.internal.utils (and src/internal in the
TS core). validateLabel had a single consumer, PlatformZones, so the
label check now lives there directly in both languages.
