+++ kinotic-core/src/test/java/org/kinotic/core/internal/utils/ZoneUtilTest.java
@@ -1,4 +1,4 @@
-package org.kinotic.core.api.service;
+package org.kinotic.core.internal.utils;
 
 import org.junit.jupiter.api.Test;
 
@@ -36,16 +36,6 @@ public class ZoneUtilTest {
         assertThrows(IllegalArgumentException.class, () -> ZoneUtil.validateZone("api/admin"));
         assertThrows(IllegalArgumentException.class, () -> ZoneUtil.validateZone("scope@api"));
         assertThrows(IllegalArgumentException.class, () -> ZoneUtil.validateZone("api#1"));
-    }
-
-    @Test
-    public void labelsMustBeSingleAndDotFree() {
-        assertDoesNotThrow(() -> ZoneUtil.validateLabel("acme-org"));
-        assertDoesNotThrow(() -> ZoneUtil.validateLabel("orders-app"));
-
-        assertThrows(IllegalArgumentException.class, () -> ZoneUtil.validateLabel("acme.org"));
-        assertThrows(IllegalArgumentException.class, () -> ZoneUtil.validateLabel("orders.*"));
-        assertThrows(IllegalArgumentException.class, () -> ZoneUtil.validateLabel("Acme-Org"));
     }
 
 }
