+++ kinotic-js/workspace/packages/core/src/api/KinoticDecorators.ts
@@ -1,7 +1,7 @@
 import { Kinotic } from '@/api/Kinotic'
 
 import { ServiceIdentifier } from '@/api/ServiceIdentifier'
-import { validateZone } from '@/api/ZoneUtil'
+import { validateZone } from '@/internal/ZoneUtil'
 
 /**
  * Decorator for registering services with the Kinotic ServiceRegistry.
