+++ kinotic-js/workspace/packages/core/src/internal/ZoneUtil.ts
@@ -6,7 +6,6 @@
 
 // DNS label rule per dot separated label, so a zone can never contain characters that would
 // break CRI parsing, dot-boundary prefix matching, or wildcard patterns
-const LABEL_PATTERN = /^[a-z0-9]([a-z0-9-]*[a-z0-9])?$/
 const ZONE_PATTERN = /^[a-z0-9]([a-z0-9-]*[a-z0-9])?(\.[a-z0-9]([a-z0-9-]*[a-z0-9])?)*$/
 
 /**
@@ -18,17 +17,5 @@ const ZONE_PATTERN = /^[a-z0-9]([a-z0-9-]*[a-z0-9])?(\.[a-z0-9]([a-z0-9-]*[a-z0-
 export function validateZone(zone: string): void {
     if (!zone || !ZONE_PATTERN.test(zone)) {
         throw new Error(`Invalid zone '${zone}'. Zones must be dot separated labels of lowercase letters, digits, and interior dashes`)
-    }
-}
-
-/**
- * Validates that the given value is a single zone label
- * @param label the label to validate
- * @throws if the label is empty or is not a single label of lowercase letters, digits, and
- *         interior dashes
- */
-export function validateLabel(label: string): void {
-    if (!label || !LABEL_PATTERN.test(label)) {
-        throw new Error(`Invalid zone label '${label}'. Labels must be lowercase letters, digits, and interior dashes`)
     }
 }
